shared/pyproject.toml
[project]
name = "prism-shared"
version = "0.1.0"
description = "Prism 共享库：DB session、JWT、认证中间件、通用模型、异常体系"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.0",
    "neo4j>=5.0",
    "redis>=5.0",
    "structlog>=24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prism_shared"]
